Support line data for geo_info

+++ db/migrate/004_create_geo_points.rb
@@ -4,9 +4,9 @@ class CreateGeoPoints < ActiveRecord::Migration
 
       t.integer :geo_info_id
 
-      t.decimal :longitude
+      t.float :longitude
 
-      t.decimal :latitude
+      t.float :latitude
 
 
     end
